[FE] fix: KakaoMap context 개별 파일로 분리
