+++ FrontEnd/src/contexts/MapContext.tsx
@@ -1,34 +1,4 @@
-// src/contexts/MapContext.tsx
-import { createContext, useContext, useState, useEffect } from "react";
-import { loadKakaoMap } from "@src/libs/kakaoMapLoader";
+import { createContext } from "react";
+import { MapContextType } from "@src/types/map";
 
-type MapContextType = {
-  map: kakao.maps.Map | null;
-  setMap: (map: kakao.maps.Map) => void;
-  isLoaded: boolean;
-};
-
-const MapContext = createContext<MapContextType | undefined>(undefined);
-
-export const MapProvider = ({ children }: { children: React.ReactNode }) => {
-  const [map, setMap] = useState<kakao.maps.Map | null>(null);
-  const [isLoaded, setIsLoaded] = useState(false);
-
-  useEffect(() => {
-    loadKakaoMap().then(() => {
-      setIsLoaded(true);
-    });
-  }, []);
-
-  return (
-    <MapContext.Provider value={{ map, setMap, isLoaded }}>
-      {children}
-    </MapContext.Provider>
-  );
-};
-
-export const useMap = () => {
-  const context = useContext(MapContext);
-  if (!context) throw new Error("useMap must be used within a MapProvider");
-  return context;
-};
+export const MapContext = createContext<MapContextType | undefined>(undefined);
