+++ FrontEnd/src/hooks/useMap.ts
@@ -1,0 +1,8 @@
+import { MapContext } from "@src/contexts/MapContext";
+import { useContext } from "react";
+
+export const useMap = () => {
+  const context = useContext(MapContext);
+  if (!context) throw new Error("useMap must be used within a MapProvider");
+  return context;
+};
